feat(frontend): rename useListContestTasks -> contest
